feat: inspection scheduled email working
